requirements.txt
PyQt6
SQLAlchemy
PyYAML
